pulling out custom tooltip into its own component
